Exnted dependencies and Connector, reduce wasm size
